backend/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.29.*
python-multipart
pydantic==2.*
python-openai
python-jobspy
openresume-parser
pdfplumber
python-docx
python-magic-bin==0.4.14
tenacity
